More updates to options

+++ stig-overcloud.sh
@@ -8,7 +8,10 @@ read -s -p "RHN password (doesn't echo):" password ; echo
 read -p "Subscritpion pool id:" pool_id
 oscap info /usr/share/xml/scap/ssg/content/ssg-rhel7-ds.xml |grep xccdf_org.ssgproject.content|sed -e 's/^[[:space:]]*//'|cut -d "_" -f4
 read -p "Select a Profile (copy and paste):" profile
+export morehard=n
 read -p "Do you want to apply addtional hardening from ssg-supplemental? (y/n)" morehard
+export wholecloud=n
+read -p "Do you want use whole cloud images? (y/n)" wholecloud
 rm -f overcloud-full* ironic-python-agent*
 echo "##########################"
 echo "# Getting Factory Images #"
@@ -48,21 +51,33 @@ echo "# Unregistering and Unsubscribing Image #"
 echo "#########################################"
 virt-customize -a overcloud-full.qcow2 --run-command 'subscription-manager remove --all'
 virt-customize -a overcloud-full.qcow2 --run-command 'subscription-manager unregister'
+if [[ $wholecloud =~ ^[Yy]$ ]]
+then
 echo "##############################"
 echo "# Creating Partitioned Image #"
 echo "##############################"
 #./whole-disk-image.py
 #mv /tmp/overcloud-full-partitioned.qcow2 ./overcloud-full.qcow2
 #sync
+fi
+
 virt-customize --selinux-relabel -a overcloud-full.qcow2
 echo "######################################"
 echo "# Uploading Hardened Image to Glance #"
 echo "######################################"
 source ~/stackrc
+
 openstack overcloud plan delete overcloud
+if [[ $wholecloud =~ ^[Yy]$ ]]
+then
 for i in $(openstack image list |grep overcloud |awk '{print $2}'); do openstack image delete $i ; done
 openstack overcloud image upload --whole-disk --os-image-name overcloud-full.qcow2
 openstack baremetal configure boot
+fi
+if [[ $wholecloud =~ ^[Nn]$ ]]
+then
+openstack overcloud image upload --update-existing  --image-path $(pwd)
+fi
 echo "#################################"
 echo "# Your Image is ready to deploy #"
 echo "#################################"
